Move DelayedRender to common components
