+++ src/components/common/DelayedRender.ts
@@ -1,0 +1,17 @@
+import { Fragment, PropsWithChildren, useEffect, useState } from 'react'
+import { h } from 'react-hyperscript-helpers'
+
+
+type DelayedRenderProps = PropsWithChildren<{
+  delay?: number
+}>
+
+export const DelayedRender = ({ children = null, delay = 1000 }: DelayedRenderProps) => {
+  const [shouldRender, setShouldRender] = useState(false)
+  useEffect(() => {
+    const timeout = setTimeout(() => setShouldRender(true), delay)
+    return () => clearTimeout(timeout)
+  }, [delay])
+
+  return shouldRender ? h(Fragment, [children]) : null
+}
